requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn==0.15.0
python-multipart==0.0.5
aiomysql==0.2.0  # Async MySQL driver
PyMySQL>=1.0.0   # Required by aiomysql
redis>=4.2.0  # Updated version that includes asyncio support
aioredis==1.3.1  # Async Redis client for Python
python-dotenv==0.19.0
aiohttp>=3.8.0   # For async HTTP requests

# Email validation
dnspython==2.1.0

# HTTP requests
requests==2.26.0
pydantic>=2.5.0

# Google OAuth and API
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.23.0

# Microsoft OAuth
msal==1.16.0

# Testing (optional)
pytest==7.0.1
pytest-mock==3.7.0
httpx==0.19.0  # For FastAPI testing

# Development tools (optional)
black==22.3.0
flake8==4.0.1

# Add to existing requirements
PyJWT==2.3.0 